refactor: move/rename Chart file
