+++ web/pages/admin/hardware-info.tsx
@@ -2,7 +2,7 @@ import { BulbOutlined, LaptopOutlined, SaveOutlined } from '@ant-design/icons';
 import { Row, Col, Typography } from 'antd';
 import React, { useEffect, useState } from 'react';
 import { fetchData, FETCH_INTERVAL, HARDWARE_STATS } from '../../utils/apis';
-import Chart from '../../components/chart';
+import Chart from '../../components/Chart/Chart';
 import StatisticItem from '../../components/statistic';
 
 // TODO: FIX TS WARNING FROM THIS.
