+++ web/pages/admin/stream-health.tsx
@@ -4,7 +4,7 @@ import { Row, Col, Typography, Space, Statistic, Card, Alert, Spin } from 'antd'
 import React, { ReactNode, useEffect, useState } from 'react';
 import { ClockCircleOutlined, WarningOutlined, WifiOutlined } from '@ant-design/icons';
 import { fetchData, FETCH_INTERVAL, API_STREAM_HEALTH_METRICS } from '../../utils/apis';
-import Chart from '../../components/chart';
+import Chart from '../../components/Chart/Chart';
 import StreamHealthOverview from '../../components/stream-health-overview';
 
 interface TimedValue {
